+++ src/atoms/heading/Heading.stories.js
@@ -1,27 +1,33 @@
 import { storiesOf } from '@storybook/vue'
 
 import AHeading from './Heading.vue'
+import { select, text } from '@storybook/addon-knobs'
+
+const info = 'Check **Knobs** tab to edit component properties dynamically, All HTML six levels of headings are available.'
+
+const levels = [1, 2, 3, 4, 5, 6]
 
 storiesOf('Atoms/Heading', module)
-  .addParameters({ info: true })
-  // @vue/component
+  .addParameters({ info })
   .add('Default', () => ({
     components: { AHeading },
+    props: {
+      levelKnobs: {
+        default: select('Heading level', levels, 1)
+      },
+      tagKnobs: {
+        default: text('Tag', null)
+      },
+      textKnobs: {
+        default: text('Text', 'Heading level 1')
+      }
+    },
     template: `
-      <a-heading :level="1">
-        Heading level 1
-      </a-heading>
-    `
-  }))
-  // @vue/component
-  .add('Custom tag', () => ({
-    components: { AHeading },
-    template: `
-      <a-heading
-        :level="4"
-        tag="span"
+      <a-heading 
+        :level="levelKnobs"
+        :tag="tagKnobs"
       >
-        Span level 4
+        {{ textKnobs }}
       </a-heading>
     `
   }))