+++ src/atoms/image/Image.vue
@@ -4,6 +4,8 @@
 import AImage from './Image.js'
 
 export default {
+  name: 'AlpacaImage',
+
   mixins: [AImage]
 }
 </script>