+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Move CatalogGridItem to atomic design, rename to ProductGridItem
+- heading: update heading class and add unit test and knobs
+- Update readme
 
 ### Fixed
 - Typo in Message.selectors.json file